odoo/addons/download_attachments/requirements.txt
Pillow==11.0.0
PyPDF2==1.26.0
PyPDF2==3.0.1
Werkzeug==2.0.2
